+++ Libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: jazarago <jazarago@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/20 10:40:09 by jazarago          #+#    #+#             */
-/*   Updated: 2024/02/27 12:35:16 by jazarago         ###   ########.fr       */
+/*   Updated: 2024/07/09 12:55:18 by jazarago         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ typedef struct s_list
 #  define OPEN_MAX 1042
 # endif
 
-int		ft_atoi(const char *str);
+long	ft_atoi(const char *str);
 void	ft_bzero(void *s, size_t n);
 void	*ft_calloc(size_t count, size_t size);
 int		ft_isalnum(int c);
