+++ stack/stack.h
@@ -6,7 +6,7 @@
 /*   By: jazarago <jazarago@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/11 17:39:30 by javi              #+#    #+#             */
-/*   Updated: 2024/06/12 11:05:13 by jazarago         ###   ########.fr       */
+/*   Updated: 2024/07/09 11:58:32 by jazarago         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 typedef struct node
 {
-	int			value;
+	long		value;
 	int			pos;
 	struct node	*next;
 }	t_Node;
@@ -26,11 +26,11 @@ typedef struct stack
 	int		size;
 }	t_Stack;
 
-t_Node	*ft_new_node(int value, int pos);
+t_Node	*ft_new_node(long value, int pos);
 void	ft_remove_node(t_Node *node);
 t_Stack	*ft_create_empty_stack(void);
 void	ft_destroy_stack(t_Stack *stack);
-int		ft_stack_push(t_Stack *stack, int value);
+int		ft_stack_push(t_Stack *stack, long value);
 void	ft_pop(t_Stack *stack);
 
 #endif
